chore: LuckyCardFactory 리팩토링
